+++ lib/schked/railtie.rb
@@ -4,6 +4,8 @@ require "rails/railtie"
 
 module Schked
   class Railtie < Rails::Railtie
+    config.schked = ActiveSupport::OrderedOptions.new
+
     class PathsConfig
       def self.call(app)
         return if Schked.config.do_not_load_root_schedule?
@@ -16,7 +18,15 @@ module Schked
       end
     end
 
+    class LivenessConfig
+      def self.call(app)
+        liveness_probe = app.config.schked.liveness_probe
+        Schked.config.liveness_probe = liveness_probe if liveness_probe
+      end
+    end
+
     initializer("schked.paths", &PathsConfig.method(:call))
+    initializer("schked.liveness", after: :load_config_initializers) { |app| LivenessConfig.call(app) }
 
     config.to_prepare do
       Schked.config.logger = ::Rails.logger unless Schked.config.logger?